requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
numpy
scikit-learn
opencv-python
pillow
torch
torchvision
clip
